+++ pkg/api/v1/types.go
@@ -1365,67 +1365,6 @@ type ReplicationControllerList struct {
 	Items []ReplicationController `json:"items"`
 }
 
-// DaemonSpec is the specification of a daemon.
-type DaemonSpec struct {
-	// Selector is a label query over pods that are managed by the daemon.
-	// Must match in order to be controlled.
-	// If empty, defaulted to labels on Pod template.
-	// More info: http://releases.k8s.io/HEAD/docs/user-guide/labels.md#label-selectors
-	Selector map[string]string `json:"selector,omitempty"`
-
-	// Template is the object that describes the pod that will be created.
-	// The Daemon will create exactly one copy of this pod on every node
-	// that matches the template's node selector (or on every node if no node
-	// selector is specified).
-	// More info: http://releases.k8s.io/HEAD/docs/user-guide/replication-controller.md#pod-template
-	Template *PodTemplateSpec `json:"template,omitempty"`
-}
-
-// DaemonStatus represents the current status of a daemon.
-type DaemonStatus struct {
-	// CurrentNumberScheduled is the number of nodes that are running exactly 1 copy of the
-	// daemon and are supposed to run the daemon.
-	CurrentNumberScheduled int `json:"currentNumberScheduled"`
-
-	// NumberMisscheduled is the number of nodes that are running the daemon, but are
-	// not supposed to run the daemon.
-	NumberMisscheduled int `json:"numberMisscheduled"`
-
-	// DesiredNumberScheduled is the total number of nodes that should be running the daemon
-	// (including nodes correctly running the daemon).
-	DesiredNumberScheduled int `json:"desiredNumberScheduled"`
-}
-
-// Daemon represents the configuration of a daemon.
-type Daemon struct {
-	TypeMeta `json:",inline"`
-	// Standard object's metadata.
-	// More info: http://releases.k8s.io/HEAD/docs/devel/api-conventions.md#metadata
-	ObjectMeta `json:"metadata,omitempty"`
-
-	// Spec defines the specification of the desired behavior of this daemon.
-	// More info: http://releases.k8s.io/HEAD/docs/devel/api-conventions.md#spec-and-status
-	Spec DaemonSpec `json:"spec,omitempty"`
-
-	// Status is the current status of this daemon. This data may be
-	// out of date by some window of time.
-	// Populated by the system.
-	// Read-only.
-	// More info: http://releases.k8s.io/HEAD/docs/devel/api-conventions.md#spec-and-status
-	Status DaemonStatus `json:"status,omitempty"`
-}
-
-// DaemonList is a collection of daemon.
-type DaemonList struct {
-	TypeMeta `json:",inline"`
-	// Standard list metadata.
-	// More info: http://releases.k8s.io/HEAD/docs/devel/api-conventions.md#metadata
-	ListMeta `json:"metadata,omitempty"`
-
-	// Items is a list of daemons.
-	Items []Daemon `json:"items"`
-}
-
 // Session Affinity Type string
 type ServiceAffinity string
 
@@ -2370,8 +2309,6 @@ const (
 	ResourceServices ResourceName = "services"
 	// ReplicationControllers, number
 	ResourceReplicationControllers ResourceName = "replicationcontrollers"
-	// Daemon, number
-	ResourceDaemon ResourceName = "daemon"
 	// ResourceQuotas, number
 	ResourceQuotas ResourceName = "resourcequotas"
 	// ResourceSecrets, number
