+++ pkg/client/unversioned/client.go
@@ -33,7 +33,6 @@ type Interface interface {
 	PodsNamespacer
 	PodTemplatesNamespacer
 	ReplicationControllersNamespacer
-	DaemonsNamespacer
 	ServicesNamespacer
 	EndpointsNamespacer
 	VersionInterface
@@ -51,10 +50,6 @@ type Interface interface {
 
 func (c *Client) ReplicationControllers(namespace string) ReplicationControllerInterface {
 	return newReplicationControllers(c, namespace)
-}
-
-func (c *Client) Daemons(namespace string) DaemonInterface {
-	return newDaemons(c, namespace)
 }
 
 func (c *Client) Nodes() NodeInterface {
