+++ pkg/client/unversioned/resource_quotas_test.go
@@ -45,7 +45,6 @@ func TestResourceQuotaCreate(t *testing.T) {
 				api.ResourcePods:                   resource.MustParse("10"),
 				api.ResourceServices:               resource.MustParse("10"),
 				api.ResourceReplicationControllers: resource.MustParse("10"),
-				api.ResourceDaemon:                 resource.MustParse("10"),
 				api.ResourceQuotas:                 resource.MustParse("10"),
 			},
 		},
@@ -78,7 +77,6 @@ func TestResourceQuotaGet(t *testing.T) {
 				api.ResourcePods:                   resource.MustParse("10"),
 				api.ResourceServices:               resource.MustParse("10"),
 				api.ResourceReplicationControllers: resource.MustParse("10"),
-				api.ResourceDaemon:                 resource.MustParse("10"),
 				api.ResourceQuotas:                 resource.MustParse("10"),
 			},
 		},
@@ -135,7 +133,6 @@ func TestResourceQuotaUpdate(t *testing.T) {
 				api.ResourcePods:                   resource.MustParse("10"),
 				api.ResourceServices:               resource.MustParse("10"),
 				api.ResourceReplicationControllers: resource.MustParse("10"),
-				api.ResourceDaemon:                 resource.MustParse("10"),
 				api.ResourceQuotas:                 resource.MustParse("10"),
 			},
 		},
@@ -163,7 +160,6 @@ func TestResourceQuotaStatusUpdate(t *testing.T) {
 				api.ResourcePods:                   resource.MustParse("10"),
 				api.ResourceServices:               resource.MustParse("10"),
 				api.ResourceReplicationControllers: resource.MustParse("10"),
-				api.ResourceDaemon:                 resource.MustParse("10"),
 				api.ResourceQuotas:                 resource.MustParse("10"),
 			},
 		},
