+++ pkg/expapi/register.go
@@ -36,6 +36,8 @@ func addKnownTypes() {
 		&Scale{},
 		&ThirdPartyResource{},
 		&ThirdPartyResourceList{},
+		&DaemonList{},
+		&Daemon{},
 	)
 }
 
@@ -47,3 +49,5 @@ func (*ReplicationControllerDummy) IsAnAPIObject()  {}
 func (*Scale) IsAnAPIObject()                       {}
 func (*ThirdPartyResource) IsAnAPIObject()          {}
 func (*ThirdPartyResourceList) IsAnAPIObject()      {}
+func (*Daemon) IsAnAPIObject()                      {}
+func (*DaemonList) IsAnAPIObject()                  {}
