+++ pkg/registry/daemon/etcd/etcd.go
@@ -18,6 +18,7 @@ package etcd
 
 import (
 	"k8s.io/kubernetes/pkg/api"
+	"k8s.io/kubernetes/pkg/expapi"
 	"k8s.io/kubernetes/pkg/fields"
 	"k8s.io/kubernetes/pkg/labels"
 	"k8s.io/kubernetes/pkg/registry/daemon"
@@ -39,10 +40,10 @@ var daemonPrefix = "/daemons"
 // NewREST returns a RESTStorage object that will work against daemons.
 func NewREST(s storage.Interface) *REST {
 	store := &etcdgeneric.Etcd{
-		NewFunc: func() runtime.Object { return &api.Daemon{} },
+		NewFunc: func() runtime.Object { return &expapi.Daemon{} },
 
 		// NewListFunc returns an object capable of storing results of an etcd list.
-		NewListFunc: func() runtime.Object { return &api.DaemonList{} },
+		NewListFunc: func() runtime.Object { return &expapi.DaemonList{} },
 		// Produces a path that etcd understands, to the root of the resource
 		// by combining the namespace in the context with the given prefix
 		KeyRootFunc: func(ctx api.Context) string {
@@ -55,7 +56,7 @@ func NewREST(s storage.Interface) *REST {
 		},
 		// Retrieve the name field of a daemon
 		ObjectNameFunc: func(obj runtime.Object) (string, error) {
-			return obj.(*api.Daemon).Name, nil
+			return obj.(*expapi.Daemon).Name, nil
 		},
 		// Used to match objects based on labels/fields for list and watch
 		PredicateFunc: func(label labels.Selector, field fields.Selector) generic.Matcher {
